pyproject.toml
[project]
name = "mpptoxls"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = ["jpype1"]
